Created CV App
Created a CV app with react. Broken the project into different
components, some for input some for the display of the CV. Still need to
organize the codebase, refactor it and style the pages properly. No
props validation yet. No input validation at all, needs to be added.
Want to add a clear all button and a print function via API.